feat: 随机生成实现，调整相关整数类型。

+++ src/components/position.rs
@@ -3,6 +3,6 @@ use specs::{Component, VecStorage};
 #[derive(Component)]
 #[storage(VecStorage)]
 pub struct Position {
-    pub x: u32,
-    pub y: u32,
+    pub x: u64,
+    pub y: u64,
 }
