+++ src/components/viewshed.rs
@@ -4,8 +4,8 @@ use specs::{Component, VecStorage};
 #[derive(Component)]
 #[storage(VecStorage)]
 pub struct Viewshed {
-    pub visible_tiles: Vec<Point2<u32>>,
-    pub range: u32,
+    pub visible_tiles: Vec<Point2<u64>>,
+    pub range: u64,
     /// flag indicate the field of view changed.
     pub invalid: bool,
 }
